hvac_collector/requirements.txt
# HVAC Energy Collector dependencies
# Install separately from the FastAPI backend:
#   cd hvac_collector && pip install -r requirements.txt

pyeconet
thinq-connect
supabase
aiohttp
requests
pandas
python-dotenv
